whisper-api/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
aiofiles
